+++ DesignToolbox/DesignToolbox/Pages/About/AboutPage.swift
@@ -106,20 +106,57 @@ struct AboutPage: View {
 
     @ViewBuilder
     private var buildView: some View {
-        // TODO: Only for debug purposes, should be displayed in another way
-        DesignToolboxCopyableText("app_about_details_appVersion" <- Bundle.main.marketingVersion, Bundle.main.marketingVersion)
-        DesignToolboxCopyableText("app_about_details_buildNumber" <- Bundle.main.buildNumber, Bundle.main.buildNumber)
-        DesignToolboxCopyableText("app_about_details_buildType" <- Bundle.main.fullBuildType, Bundle.main.fullBuildType)
-        if let buildDetails = Bundle.main.buildDetails {
-            DesignToolboxCopyableText("app_about_details_githubBuildDetails" <- buildDetails,
-                                      buildDetails.leadingColumnFragment)
-        }
-        if let sdkVersion = Bundle.main.sdkVersion, !sdkVersion.isEmpty {
-            DesignToolboxCopyableText("app_about_details_sdkVersion" <- sdkVersion, sdkVersion)
+
+        if Bundle.main.fullBuildType == "stable" {
+            OpenableText("app_about_details_appVersion" <- Bundle.main.marketingVersion, anchor: Bundle.main.marketingVersion, type: .githubRelease)
+                .modifier(CopyableTextViewModifier(Bundle.main.marketingVersion))
+        } else {
+            Text("app_about_details_appVersion" <- Bundle.main.marketingVersion)
+                .modifier(CopyableTextViewModifier(Bundle.main.marketingVersion))
         }
 
-        DesignToolboxCopyableText("app_about_details_themeCoreVersion" <- OUDSVersions.themeCoreVersion, OUDSVersions.themeCoreVersion)
-        DesignToolboxCopyableText("app_about_details_themeOrangeVersion" <- OUDSVersions.themeOrangeVersion, OUDSVersions.themeOrangeVersion)
+        Text("app_about_details_buildNumber" <- Bundle.main.buildNumber)
+            .modifier(CopyableTextViewModifier(Bundle.main.buildNumber))
+
+        OpenableText("app_about_details_buildType" <- Bundle.main.fullBuildType, anchor: Bundle.main.fullBuildType, type: .githubBuild)
+            .modifier(CopyableTextViewModifier(Bundle.main.fullBuildType))
+
+        if let buildDetails = Bundle.main.buildDetails {
+            OpenableText("app_about_details_githubBuildDetails" <- buildDetails, anchor: buildDetails, type: .githubIssue)
+                .modifier(CopyableTextViewModifier(buildDetails.leadingColumnFragment))
+        }
+
+        if let sdkVersion = Bundle.main.sdkVersion, !sdkVersion.isEmpty {
+            OpenableText("app_about_details_sdkVersion" <- sdkVersion, anchor: sdkVersion, type: .githubVersion)
+                .modifier(CopyableTextViewModifier(sdkVersion))
+        }
+
+        Text("app_about_details_themeCoreVersion" <- OUDSVersions.themeCoreVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeCoreVersion))
+
+        Text("app_about_details_themeOrangeCoreVersion" <- OUDSVersions.themeOrangeCoreVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeOrangeCoreVersion))
+
+        Text("app_about_details_themeSoshCoreVersion" <- OUDSVersions.themeSoshCoreVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeSoshCoreVersion))
+
+        Text("app_about_details_themeWireframeCoreVersion" <- OUDSVersions.themeWireframeCoreVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeWireframeCoreVersion))
+
+        Text("app_about_details_themeOrangeBrandVersion" <- OUDSVersions.themeOrangeBrandVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeOrangeBrandVersion))
+
+        Text("app_about_details_themeSoshBrandVersion" <- OUDSVersions.themeSoshBrandVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeSoshBrandVersion))
+
+        Text("app_about_details_themeOrangeBusinessToolsBrandVersion" <- OUDSVersions.themeOrangeBusinessToolsBrandVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeOrangeBusinessToolsBrandVersion))
+
+        Text("app_about_details_themeOrangeInverseBrandVersion" <- OUDSVersions.themeOrangeInverseBrandVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeOrangeInverseBrandVersion))
+
+        Text("app_about_details_themeWireframeBrandVersion" <- OUDSVersions.themeWireframeBrandVersion)
+            .modifier(CopyableTextViewModifier(OUDSVersions.themeWireframeBrandVersion))
     }
 
     @ViewBuilder