+++ DesignToolbox/DesignToolbox/Pages/Components/ComponentsPage.swift
@@ -18,9 +18,11 @@ import SwiftUI
 struct ComponentsPage: View {
 
     let componentElements: [DesignToolboxElement] = [
+        BadgeElement(),
         ButtonElement(),
         ColoredSurfaceElement(),
         CheckboxElements(),
+        ChipElements(),
         DividerElements(),
         LinkElement(),
         RadioElements(),