+++ DesignToolbox/DesignToolbox/Pages/Tokens/Color/ColorTokenPage.swift
@@ -31,7 +31,9 @@ struct ColorTokenPage: View {
             Section { illustrationForAlways() } header: { header("Always") }
             Section { illustrationForBackground() } header: { header("Background") }
             Section { illustrationForBorder() } header: { header("Border") }
-            Section { illustrationForChart() } header: { header("Chart") }
+            if theme.colorCharts != nil { // Some themes like Sosh do not have color charts
+                Section { illustrationForChart() } header: { header("Chart") }
+            }
             Section { illustrationForContent() } header: { header("Content") }
             Section { illustrationForDecorative() } header: { header("Decorative") }
             Section { illustrationForOpacity() } header: { header("Opacity") }
@@ -51,7 +53,9 @@ struct ColorTokenPage: View {
     private func illustrationForBackground() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Background.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -59,7 +63,9 @@ struct ColorTokenPage: View {
     private func illustrationForAction() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Action.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -67,7 +73,9 @@ struct ColorTokenPage: View {
     private func illustrationForAlways() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Always.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).isForbiddenValueColor() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -75,7 +83,9 @@ struct ColorTokenPage: View {
     private func illustrationForChart() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Chart.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -83,7 +93,9 @@ struct ColorTokenPage: View {
     private func illustrationForBorder() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Border.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -91,7 +103,9 @@ struct ColorTokenPage: View {
     private func illustrationForContent() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Content.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -99,7 +113,9 @@ struct ColorTokenPage: View {
     private func illustrationForDecorative() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Decorative.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).isForbiddenValueColor() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -107,7 +123,9 @@ struct ColorTokenPage: View {
     private func illustrationForOverlay() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Overlay.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -115,7 +133,9 @@ struct ColorTokenPage: View {
     private func illustrationForSurface() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Surface.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -123,7 +143,9 @@ struct ColorTokenPage: View {
     private func illustrationForOpacity() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Opacity.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -131,7 +153,9 @@ struct ColorTokenPage: View {
     private func illustrationForRepository() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Repository.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).isForbiddenValueColor() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -159,7 +183,6 @@ struct ColorTokenPage: View {
 
         var body: some View {
             let colorRawToken = colorScheme == .dark ? lightValue : darkValue
-
             DesignToolboxTokenIllustration(tokenName: name, tokenValue: colorRawToken) {
                 Rectangle()
                     .fill(colorRawToken.color)