+++ CHANGELOG.md
@@ -4,6 +4,69 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/0.16.0...0.17.0) - 2025-07-24
+
+### Added
+
+- [Library] Wireframe theme (Orange-OpenSource/ouds-ios#669)
+- [Library] Text input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Text area component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Select input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Quantity input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Tag input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Pin code input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Color charts for Orange Business Tools theme
+- [Library] Orange Inverse theme (Orange-OpenSource/ouds-ios#828)
+- [Library] Orange Business Tools theme (Orange-OpenSource/ouds-ios#821)
+- [DesignToolbox] GitHub references in about page (Orange-OpenSource/ouds-ios#852)
+- [DesignToolbox] Alert killing or not the app when theme changed (if toggled) (Orange-OpenSource/ouds-ios#850)
+- [Library] Chip components (Orange-OpenSource/ouds-ios#407)
+- [Library] Add `badge` components (Orange-OpenSource/ouds-ios#514) (Orange-OpenSource/ouds-ios#847)
+
+### Changed
+
+- [Library] Color semantic tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
+- [Library] Border semantic tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
+- [Library] Text input component tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
+- [Library] Button component tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
+- [Library] Control item component (Orange-OpenSource/ouds-ios#875)
+- [Library] Opacity semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Color repository semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Color mode semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Space semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Tag component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Link component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Chip component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Button component (Orange-OpenSource/ouds-ios#875)
+- [Library] Button component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Bullet list component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Split color charts semantic tokens to dedicated provider (Orange-OpenSource/ouds-ios#856)
+- [DesignToolbox] Update `fastlane` gem from v2.227.2 to v2.228.0
+- [DesignToolbox] Update `SwiftFormat/CLI` pod from v0.56.4 to v0.57.2
+- [Library] Swift package `swift-docc-plugin` from v1.4.3 to v1.4.5
+- [Library] Update `fastlane` gem from v2.227.2 to v2.228.0
+- [Library] Swift package `SwifFormat` from v0.56.4 to v0.57.2
+- [DesignToolbox] Update the position of the design component version (Orange-OpenSource/ouds-ios#848)
+- [DesignToolbox] Replace components illustrations by the components themselves (Orange-OpenSource/ouds-ios#820)
+- [DesignToolbox] Update `badge` component screen (Orange-OpenSource/ouds-ios#840)
+- [DesignToolbox] Available color modes and surfaces by theme for `colored surface` component
+- [Library] Color mode semantic tokens (tokens library v1.1.0)
+- [Library] Color multiple semantic tokens (tokens library v1.1.0)
+
+### Removed
+
+- [Library] List item component tokens
+- [Library] Input text component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Select component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- [Library] Orange theme font family raw tokens
+
+### Fixed
+
+- [DesignToolbox] Fix default indicator position in `switch control item` (Orange-OpenSource/ouds-ios#845)
+- [Library] Update `controlItem` component to use `controlItemBorderRadius*` tokens (Orange-OpenSource/ouds-ios#777)
+- [Library] Update `switch` component to use `switchBorderRadius*` tokens (Orange-OpenSource/ouds-ios#780)
+- [Library] Sosh theme (Orange-OpenSource/ouds-ios#665)
+
 ## [0.16.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/0.15.0...0.16.0) - 2025-07-07
 
 ### Added
@@ -53,6 +116,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- [Library] Divider in a ControlItem should not increase the size of this component (Orange-OpenSource/ouds-ios#525)
 - [Library] Background color in pressed and hover states for Radio button and Checkbox components (Orange-OpenSource/ouds-ios#720)
 - [DesignToolbox] Update background color in component illustration view (Orange-OpenSource/ouds-ios#723)
 - [DesignToolbox] In radio item page icons never flipped when flip option toggled (Orange-OpenSource/ouds-ios#675)